feat: web loader parameters
