+++ apps/cloud/src/app/@shared/common/select/select.component.html
@@ -1,7 +1,7 @@
 <div tabindex="0"
   #menuTrigger
   #trigger="cdkMenuTriggerFor"
-  class="select-trigger flex items-center rounded-lg cursor-pointer hover:bg-gray-200"
+  class="select-trigger grow flex items-center rounded-lg cursor-pointer hover:bg-gray-200"
   [cdkMenuTriggerFor]="menu"
   [cdkMenuTriggerData]="{ trigger: trigger }"
 >
