+++ packages/server-ai/src/rag-web/queries/handlers/get-options.handler.ts
@@ -22,28 +22,71 @@ export class GetRagWebOptionsHandler implements IQueryHandler<GetRagWebOptionsQu
 				{
 					name: 'url',
 					type: ParameterTypeEnum.STRING,
+					label: {
+						zh_Hans: '链接',
+					},
 					placeholder: {
 						zh_Hans: '链接'
+					},
+					description: {
+						zh_Hans: '如何抓取链接',
 					}
 				},
 				{
 					name: 'mode',
 					type: ParameterTypeEnum.SELECT,
+					label: {
+						zh_Hans: '模式',
+					},
+					placeholder: {
+						zh_Hans: '选择抓取模式'
+					},
+					description: {
+						zh_Hans: '如何抓取模式',
+					},
 					options: [
 						{
 							value: 'scrape',
 							label: {
-								en_US: 'scrape single urls'
+								en_US: 'Scrape single urls'
 							}
 						},
 						{
 							value: 'crawl',
 							label: {
-								en_US: 'crawl all accessible subpages'
+								en_US: 'Crawl all accessible subpages'
 							}
 						}
 					]
-				}
+				},
+				{
+					when: {mode: ['crawl']},
+					name: 'ignoreSitemap',
+					type: ParameterTypeEnum.BOOLEAN,
+					label: {
+						zh_Hans: '忽略 Sitemap',
+					},
+					placeholder: {
+						zh_Hans: '忽略 Sitemap'
+					},
+					description: {
+						en_US: 'Ignore the website sitemap when crawling',
+					}
+				},
+				{
+					when: {mode: ['crawl']},
+					name: 'limit',
+					type: ParameterTypeEnum.NUMBER,
+					label: {
+						zh_Hans: 'Maximum num of pages',
+					},
+					placeholder: {
+						zh_Hans: 'Maximum num of pages'
+					},
+					description: {
+						en_US: 'Maximum number of pages to crawl. Default limit is 10000.',
+					}
+				},
 			]
 		} as TKDocumentWebSchema
 	}
